requirements.txt
SQLAlchemy
pyodbc
PyJWT
Flask
python-dotenv
flask-cors
pillow
